+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **[NEW]** `ButtonGroup` component
 - **[FIX]** Do not render empty itinerary steps when its title is a string made of whitespaces.
+- **[FIX]** `Menu` component configures correctly its items onClick handlers.
+- [...]
 
 # v0.23.0 (28/02/2019)